refact: rename app to DummyHTTP
The app does not sit under the HTTP namespace anymore (that was bugging
me) and has a unique name (over the web, github, etc.).

Also this makes the binary not namespace and allows easy Heroku
deployment, as the buildpack uses the app's name as the src/app_name.cr
main file to compile.